+++ CHANGELOG.md
@@ -11,6 +11,12 @@ CHANGELOG
       fzf --preview='fzf-preview.sh {}'
       ```
 - (Experimental) Sixel and Kitty image support now also available on Windows
+- HTTP server can be configured to accept remote connections
+  ```sh
+  # FZF_API_KEY is required for a non-localhost listen address
+  export FZF_API_KEY="$(head -c 32 /dev/urandom | base64)"
+  fzf --listen 0.0.0.0:6266
+  ```
 - Bug fixes
 
 0.43.0